move data_objects.py to datastore module
