+++ mediachain/transactor.py
@@ -1,7 +1,8 @@
 from grpc.beta import implementations
-from mediachain.proto import Transactor_pb2
-from mediachain.data_objects import Artefact, Entity, ChainCell, \
+
+from mediachain.datastore.data_objects import Artefact, Entity, ChainCell, \
     MultihashReference
+from mediachain.proto import Transactor_pb2
 
 TIMEOUT_SECS = 120
